refactor: move 11 retrieval modules into retrieval/
