+++ helix_context/retrieval/expand.py
@@ -36,8 +36,8 @@ import logging
 import sqlite3
 from typing import Any, Dict, List, Optional, Tuple
 
-from . import session_delivery as _session_delivery
-from .schemas import Gene
+from .. import session_delivery as _session_delivery
+from ..schemas import Gene
 
 log = logging.getLogger("helix.expand")
